moved fileops into metrics.go, fixed some review comments
